+++ include/materials.hpp
@@ -27,11 +27,12 @@ struct CheckeredPigment;
 
 class BRDF;
 class DiffusiveBRDF;
+class SpecularBRDF;
 
 class Material;
 
 //-------------------------------------------------------------------------------------------------------------
-// -----------PIGMENT STRUCT------------------
+// -----------PIGMENT ABSTRACT STRUCT ------------------
 // ------------------------------------------------------------------------------------------------------------
 
 /// @brief abstract functor that associates a Color to a Vec2d
@@ -40,6 +41,7 @@ struct Pigment {
   //------------Methods-----------
   virtual Color operator()(Vec2d uv) const = 0;
 };
+
 
 //-------------------------------------------------------------------------
 //---------------------- UNIFORM PIGMENT STRUCT ------------------
@@ -71,7 +73,7 @@ struct CheckeredPigment : public Pigment {
 
   //-------Properties--------
   Color color1, color2;
-  int n_intervals; // number of intervals into which the range [0, 1] is divided for u and v
+  int n_intervals; // number of intervals into which the range [0, 1] is divided for u and v cohordinates
 
   //-----------Constructor-----------
   CheckeredPigment(Color color1 = Color(), Color color2 = Color(), int n_intervals = 10)
@@ -81,12 +83,13 @@ struct CheckeredPigment : public Pigment {
   virtual Color operator()(Vec2d uv) const {
     float subinterval = 1.f / n_intervals; // length of u and v subintervals
 
+    // get the column and row of the subinterval in which the (u, v) coordinates fall
     int col = std::floor(uv.u / subinterval);
     int row = std::floor(uv.v / subinterval);
-    if ((col + row) % 2 == 0) { // top-left rectangle has color1
+    if ((col + row) % 2 == 0) { // entry has color1 if col and row have same parity
       return color1;
     } else {
-      return color2;
+      return color2; // it has color 2 otherwise
     }
   };
 };
@@ -146,19 +149,22 @@ public:
   BRDF(std::shared_ptr<Pigment> pigment) : pigment(pigment) {};
 
   //------------Methods-----------
-  /// @brief returns a Color corresponding to the BRDF integrated over r, g, b bands
+
+  /// @brief returns the BRDF integrated over r, g, b bands, that is 3 scalar values as a color object
   /// @param normal at hitting point
   /// @param incident direction
   /// @param outgoing direction
-  /// @ (u, v) coordinates
+  /// @param uv coordinates of the point on the surface
   virtual Color eval(Normal normal, Vec in_dir, Vec out_dir, Vec2d uv) const = 0;
+  // TODO the virtual method implemented in BRDF parent class passes arguments by direct value, wouldn't it be better to
+  // pass them by const reference?
 
   /// @brief scatters ray in random direction using BRDF-based importance sampling
-  /// @param used to generate random numbers
-  /// @param direction of the incolming ray
-  /// @param point where the incoming ray hits the surface
+  /// @param pcg used to generate random numbers for importance sampling MC
+  /// @param direction of the incoming ray
+  /// @param point world point where the incoming ray hits the surface
   /// @param normal to the surface at that point
-  /// @param depth value for the newly scattered ray
+  /// @param depth value for the newly scattered ray, counting how many times it has been reflected
   virtual Ray scatter_ray(std::shared_ptr<PCG> pcg, Vec incoming_dir, Point intersection_point, Normal normal,
                           int depth) const = 0;
 };
@@ -170,21 +176,23 @@ public:
 class DiffusiveBRDF : public BRDF {
 public:
   //-------Properties--------
-  float reflectance; // reflectance of the object
-  // QUESTION isn't reflactance already encoded in the pigment (three reflectances, one for each band)???
 
   //-----------Constructor-----------
-  /// @param reflectance of the object
   /// @param pigment of the object
-  DiffusiveBRDF(std::shared_ptr<Pigment> pigment = nullptr, float reflectance = 1.f) : BRDF(pigment), reflectance(reflectance) {
+  DiffusiveBRDF(std::shared_ptr<Pigment> pigment = nullptr) : BRDF(pigment) {
     if (!this->pigment) {
-      this->pigment = std::make_shared<UniformPigment>();
+      this->pigment = std::make_shared<UniformPigment>(); // if no pigment is provided, set uniform pigment (black)
     }
   };
 
   //------------Methods-----------
+
+  /// @brief evaluate the BRDF at the given point, by definition diffusive BRDF is just the color/pigment divided by pi
+  // in fact this method will not be used since BRDF simplifies in importance sampling for MC pathtracing
   Color eval(Normal normal, Vec in_dir, Vec out_dir, Vec2d uv) const override {
-    return (*pigment)(uv)*reflectance * (1.f / std::numbers::pi);
+    // TODO the virtual method implemented in BRDF parent class passes arguments by direct value, wouldn't it be better to
+    // pass them by const reference?
+    return (*pigment)(uv) * (1.f / std::numbers::pi); // dereference pigment pointer, get color at uv coordinates, divide by pi
   }
 
   Ray scatter_ray(std::shared_ptr<PCG> pcg, Vec incoming_dir, Point intersection_point, Normal normal, int depth) const override {
@@ -193,9 +201,8 @@ public:
     auto [theta, phi] = pcg->random_phong(1); // uniform BRDF makes the integrand of the rendering equation proportional to
                                               // cos(theta), hence we perform importance sampling using Phong n=1 distribution
     Vec outgoing_dir{onb.e1 * std::sin(theta) * std::cos(phi) + onb.e2 * std::sin(theta) * std::sin(phi) +
-                     onb.e3 * std::cos(theta)};
+                     onb.e3 * std::cos(theta)}; // get outgoing direction from the local ONB basis
 
-    // QUESTION why should tmin be bigger than usual? see lab 11 slide 13
     return Ray(intersection_point, outgoing_dir, 1.e-3f, infinite, depth);
   };
 };
@@ -257,23 +264,27 @@ public:
 // -----------MATERIAL CLASS------------------
 // ------------------------------------------------------------------------------------------------------------
 
-/// @brief light emitting and reflective properties of an object as a function of (u, v)
+/// @brief light emissive and reflective properties of a shape object as a function of (u, v) coordinates on the surface
 class Material {
 public:
   //-------Properties--------
   std::shared_ptr<BRDF> brdf;
-  std::shared_ptr<Pigment> emitted_radiance;
+  std::shared_ptr<Pigment> emitted_radiance; // Pigment that describes the emitted radiance of the material, if any
 
   //-----------Constructors-----------
   Material(std::shared_ptr<BRDF> brdf = nullptr, std::shared_ptr<Pigment> emitted_radiance = nullptr)
       : brdf(brdf), emitted_radiance(emitted_radiance) {
     if (!this->brdf) {
-      this->brdf = std::make_shared<DiffusiveBRDF>();
+      this->brdf = std::make_shared<DiffusiveBRDF>(); // if no BRDF is provided, set diffusive BRDF with uniform pigment black
     }
     if (!this->emitted_radiance) {
-      this->emitted_radiance = std::make_shared<UniformPigment>();
+      this->emitted_radiance =
+          std::make_shared<UniformPigment>(); // if no emitted radiance is provided, set uniform pigment black
     }
   }
+
+  // Costructor with only emitted radiance
+  Material(std::shared_ptr<Pigment> emitted_radiance) : Material(nullptr, emitted_radiance) {}
 
   //------------Methods-----------
 };