+++ include/renderers.hpp
@@ -18,9 +18,15 @@
 // ----------------------------------------------------------------------------------------
 
 class Renderer;
+
+class OnOffTracer;
 class FlatTracer;
 class PointLightTracer;
 class PathTracer;
+
+// ------------------------------------------------------------------------------------------------------------
+// --------------- RENDERER ABSTRACT CLASS -------------------------------------------------
+// ------------------------------------------------------------------------------------------------------------
 
 /// @brief abstract functor that associates a Color to a Ray
 class Renderer {
@@ -32,11 +38,41 @@ public:
   //-----------Constructors-----------
   Renderer(std::shared_ptr<World> world, Color background_color = Color()) : world(world), background_color(background_color) {};
 
+
   //------------Methods-----------
   virtual Color operator()(Ray ray) const = 0;
 };
 
-/// @brief functor performing flat tracing, i. e. returning for each ray the color of the closest object to be hit
+// ------------------------------------------------------------------------------------------------------------
+// --------------- ON_OFF RENDERER -------------------------------------------------
+// ------------------------------------------------------------------------------------------------------------
+
+/// @brief functor performing ON/OFF tracing: returns white if the ray hits any object, black otherwise
+class OnOffTracer : public Renderer {
+public:
+  //------------Properties-----------
+
+  //-----------Constructors-----------
+  /// Constructor with parameters
+  /// @param world to render
+  OnOffTracer(std::shared_ptr<World> world) : Renderer(world) {};
+
+  //------------Methods-----------
+  virtual Color operator()(Ray ray) const {
+    if (world->on_off_ray_intersection(
+            ray)) { // use ad hoc implemented on_off_ray_intersection method to stop looping over objects as soon as one is hit
+      return WHITE; // return white if any object is hit
+    } else {
+      return Color(); // return black if no object is hit
+    }
+  }
+};
+
+//----------------------------------------------------------------------------------------------------------------
+// --------------- FLAT RENDERER -------------------------------------------------
+//----------------------------------------------------------------------------------------------------------------
+
+/// @brief functor performing flat tracing, i. e. returning for each ray just the plain color of the closest object hit
 class FlatTracer : public Renderer {
 public:
   //------------Properties-----------
@@ -49,16 +85,20 @@ public:
 
   //------------Methods-----------
   virtual Color operator()(Ray ray) const {
-    // Save the colosest hit or return background Color if no object gets hit
-    std::optional<HitRecord> hit = world->ray_intersection(ray);
+    std::optional<HitRecord> hit = world->ray_intersection(ray); // save the closest hit (if any)
     if (!hit) {
       return background_color;
     }
-
-    // Return Color of the hit object
-    return (*hit->shape->material->brdf->pigment)(hit->surface_point);
+    // Return the color of the closest object hit (if any), both the brdf pigment and the emitted radiance 
+    return (*(hit->shape->material->brdf->pigment))(hit->surface_point) +
+           (*(hit->shape->material->emitted_radiance))(hit->surface_point);
   };
 };
+
+
+// ------------------------------------------------------------------------------------------------------------
+// --------------- POINT LIGHT TRACER -------------------------------------------------
+// ------------------------------------------------------------------------------------------------------------
 
 /// @brief functor performing point light tracing (see PR #11)
 class PointLightTracer : public Renderer {
@@ -93,11 +133,8 @@ public:
       hit_material = hit->shape->material;
       brdf = hit_material->brdf;
 
-      auto specular = false; // NOTE delete this line and restore the next one when implementation of SpecularBRDF is ready
-      // auto specular = std::dynamic_pointer_cast<SpecularBRDF>(brdf); // returns nullptr if brdf is not a pointer to
-      // an object of the derived class SpecularBRDF
-      // 3. In case you hit an object with SpecularBRDF, scatter a new Ray from the hit point in the direction given by
-      // the reflection law; otherwise go ahead with color evaluation
+      auto specular = std::dynamic_pointer_cast<SpecularBRDF>(brdf); // returns nullptr if brdf is not a pointer to an object of the derived class SpecularBRDF
+      // 3. In case you hit an object with SpecularBRDF, scatter a new Ray from the hit point in the direction given by the reflection law; otherwise go ahead with color evaluation
       if (!specular) {
         break;
       }
@@ -129,6 +166,11 @@ public:
   };
 };
 
+// ------------------------------------------------------------------------------------------------------------
+// --------------- PATH TRACER -------------------------------------------------
+// ------------------------------------------------------------------------------------------------------------
+
+
 /// @brief functor performing the path tracing algorithm
 /// @details importance sampling in MC integration based on the scatter_ray method of the BRDF
 class PathTracer : public Renderer {
@@ -136,21 +178,20 @@ public:
   //-------Properties--------
   std::shared_ptr<PCG> pcg; // random number generator
   int n_rays;               // number of rays recursively scattered
-  int russian_roulette_lim; // maximum ray depth before russian roulette kicks in
+  int russian_roulette_lim; // ray depth reached before russian roulette starts applying
   int max_depth;            // maximum ray depth
 
   //-----------Constructors-----------
-  // TODO choose reasonable value for n_rays
 
   /// @brief constructor with parameters
   /// @param world to render
-  /// @param pcg
+  /// @param pcg random number generator
   /// @param number of rays recursively scattered
-  /// @param ray depth before russian roulette kicks in
+  /// @param ray depth reached before russian roulette starts applying
   /// @param maximum ray depth
   /// @param background color
-  PathTracer(std::shared_ptr<World> world, std::shared_ptr<PCG> pcg = nullptr, int n_rays = 100, int russian_roulette_lim = 2,
-             int max_depth = 2, Color background = Color())
+  PathTracer(std::shared_ptr<World> world, std::shared_ptr<PCG> pcg = nullptr, int n_rays = 10, int russian_roulette_lim = 3,
+             int max_depth = 5, Color background = Color())
       : Renderer(world, background), pcg(pcg), n_rays(n_rays), russian_roulette_lim(russian_roulette_lim), max_depth(max_depth) {
     if (!this->pcg) {
       pcg = std::make_shared<PCG>();
@@ -164,7 +205,7 @@ public:
       return Color();
     }
 
-    // 2. Return background Color if no object gets hit
+    // 2. Get closest intersection of ray with world objects or return background Color if no object is hit
     std::optional<HitRecord> hit = world->ray_intersection(ray);
     if (!hit) {
       return background_color;
@@ -172,17 +213,20 @@ public:
 
     // 3. Unpack hit
     std::shared_ptr<Material> hit_material = hit->shape->material;
-    Color reflected_color = (*hit_material->brdf->pigment)(hit->surface_point);
-    Color emitted_radiance = (*hit_material->emitted_radiance)(hit->surface_point);
+    Color reflected_color = (*(hit_material->brdf->pigment))(hit->surface_point); // both pigment and emitted_radiance are
+                                                                                  // pointers
+    Color emitted_radiance = (*(hit_material->emitted_radiance))(hit->surface_point);
 
-    // 4. Apply russian roulette: decide whether to scatter new rays and renormalize the BRDF to compesate for possible
-    // truncations
+    // 4. Apply russian roulette: decide whether to scatter new rays and rescale the BRDF to compesate for possible
+    // truncations and get unbiased expected value
     float hit_lum = std::max({reflected_color.r, reflected_color.g, reflected_color.b});
     if (ray.depth > russian_roulette_lim) {
       float q = std::max(1.f - hit_lum, 0.05f);
-      if (pcg->random_float() >
-          q) { // stop with higher probability if the hit point has low reflactance: this improves efficiency without
-               // penalizing variance too much. Keep a finite stopiing probability even if hit_lum is close to 1
+      // according to Shirley & Morley, use max reflectance of hit color as Russian roulette probability
+      if (pcg->random_float() > q) {
+        // stop with higher probability if the hit point has low reflactance: this improves efficiency without increasing variance
+        // too much Keep a finite stopping probability 0.05 even if hit_lum is close to 1
+        reflected_color = reflected_color * (1.f / (1.f - q));
       } else {
         return emitted_radiance;
       }
@@ -190,27 +234,26 @@ public:
 
     // 5. Calculate reflected radiance recursively by: a) scattering rays in random directions according to the BRDF; b)
     // averaging the radiance from corresponding directions; c) multiplying by reflected color
-    // NOTE the algorithm is correct for a diffusiveBRDF provided reflected_color is rho_d for the three bands. This is
+    // Note that the algorithm is correct for a diffusiveBRDF provided reflected_color is rho_d for the three bands. This is
     // beacuse the normalization of the diffusiveBRDF exactly cancels out the normalization of the n=1 Phong
     // distribution. A further multiplicative factor should probably be supplied for other BRDFs: remember to do the
     // maths ad check.
     Color cum_radiance = Color();
-    if (hit_lum > 0.f) { // proceed with recursion only if reflection is possible
+    if (hit_lum > 0.f) { // proceed with recursion only if hit object is not perfectly absorbing (i.e. hit_lum==0)
       for (int i_ray = 0; i_ray < n_rays; i_ray++) {
         Ray new_ray = hit_material->brdf->scatter_ray(
             pcg, ray.direction, hit->world_point, hit->normal,
             ray.depth + 1); // Scatter a new ray by sampling the solid angle distribution proportional to the BRDF
-        cum_radiance = cum_radiance + (*this)(new_ray);
+        cum_radiance =
+            cum_radiance +
+            (*this)(new_ray); // dereference pointer `this' to the PathTracer class object and recurively call its operator()
       }
     }
     cum_radiance = cum_radiance * (1.f / n_rays) * reflected_color;
 
-    // 6. Add emitted radiance
+    // 6. Add emitted radiance specific of the object
     cum_radiance = cum_radiance + emitted_radiance;
 
     return cum_radiance;
   }
 };
-
-// TODO on/off tracing is currently implemented as a method of the class World and as such it requires a lambda wrapping
-// inside the main, which is not very elegant. Consider moving it here a class derived from Renderer.