+++ src/raytracer.cpp
@@ -7,21 +7,33 @@
 #include <fstream>
 #include <iostream>
 
-std::unique_ptr<HdrImage> make_demo_image(bool orthogonal, int width, int height, const Transformation &obs_transformation,
-                                          int samples_per_pixel_edge);
+
+//---------------------------------------------
+//---------- FORWARD DECLARATIONS ----------
+//---------------------------------------------
+// generate demo image with on off rendering
+std::unique_ptr<HdrImage> make_demo_image_onoff(bool orthogonal, int width, int height, float distance, const Transformation &obs_transformation, int samples_per_pixel_edge);
+// generate demo image with Monte Carlo path tracing rendering
+std::unique_ptr<HdrImage> make_demo_image_path(bool orthogonal, int width, int height, float distance, const Transformation &obs_transformation, int samples_per_pixel_edge);
+
+
+//-------------------------------------------------------------------
+//---------- MAIN FUNCTION ----------
+//-------------------------------------------------------------------
+
 
 int main(int argc, char **argv) {
-  CLI::App app{"Raytracer"};    // Define the main App
+  CLI::App app{"Raytracer"};    // Define the main CLI11 App
   argv = app.ensure_utf8(argv); // Ensure utf8 standard
 
-  // The main App requires one subcommand from the command line: the user has to choose between pfm2png and demo mode
+  // The main App requires exactly one subcommand from the command line: the user has to choose between pfm2png and demo mode
   app.require_subcommand(1);
 
   // -----------------------------------------------------------
   // Parameters that are common to at least two of the modes
   // -----------------------------------------------------------
 
-  // Default values of gamma and alpha (may be overwritten in pfm2png mode)
+  // Default values of gamma and alpha for HDR to LDR image conversion (may be overwritten in pfm2png mode)
   float gamma = 2.2f;
   float alpha = 0.18f;
 
@@ -39,38 +51,53 @@ int main(int argc, char **argv) {
   // Command line parsing for demo mode
   // -----------------------------------------------------------
 
+  // Add a subcommand for the demo mode
   auto demo_subc =
       app.add_subcommand("demo", "Run demo rendering and save PFM and PNG files"); // Returns a pointer to an App object
 
-  // Parse image width and height (# pixels)
-  demo_subc->add_option("--width", width, "Specify image width")->check(CLI::PositiveNumber)->default_val(1280);
-  demo_subc->add_option("--height", height, "Specify image height")->check(CLI::PositiveNumber)->default_val(960);
+  // Add option for on_off tracing or path tracing rendering
+  // Default is on/off tracing
+  std::string mode_str = "onoff";
+  demo_subc->add_option("-m,--mode", mode_str, "Rendering mode: on/off tracing (default) or path tracing")
+      ->check(CLI::IsMember({"onoff", "path"}))
+      ->default_val("onoff");
 
-  // Choose between perspective and orthogonal projection
+  // Add options for image width and height (# pixels) and provide description in the help output
+  // Default values are 1280x960
+  demo_subc->add_option("--width", width, "Specify image width")
+      ->check(CLI::PositiveNumber)
+      ->default_val(1280); // reject negative values
+  demo_subc->add_option("--height", height, "Specify image height")
+      ->check(CLI::PositiveNumber)
+      ->default_val(960); // reject negative values
+
+
+  // Add option for perspective or orthogonal projection
   bool orthogonal = false;
   auto orthogonal_flag = demo_subc->add_flag("--orthogonal", orthogonal, "Use orthogonal projection (default is perspective)");
 
-  // Parse output file name
+  // Add option for output file name
+  // Default value is "demo"
   demo_subc->add_option("-o,--output-file", output_file_name, "Insert name of the output PNG file")->default_val("demo");
 
-  // Parse observer transformation: composition of a translation along -VEC_X and rotation around the scene (endcoded in
-  // distance, angle phi and angle theta). Default position: Origin - VEC_X.
+  // Add option for observer transformation: rotation around the scene and camera-screen distance (only for perspective camera). Angles phi and theta: longitude and colatitude (theta=0 -> north pole). Default position along the negative x direction: theta = 90, phi = 180.
   float distance = 1.f;
   float theta = std::numbers::pi / 2.f;
-  float phi = 0.f;
+  float phi = std::numbers::pi;
 
   demo_subc->add_option("-d,--distance", distance, "Specify observer's distance")->excludes(orthogonal_flag)->default_val(1.f);
 
   demo_subc
       ->add_option_function<float>(
-          "--theta-deg", [&theta](const float &theta_deg) { theta = theta_deg / 180.f * std::numbers::pi; },
-          "Specify observer's angle theta")
+          "--theta-deg", [&theta](const float &theta_deg) { theta = (theta_deg / 180.f) * std::numbers::pi; },
+          "Specify observer's colatitude angle theta (0 degree is north pole)")
       ->default_val(90.f);
 
   demo_subc
       ->add_option_function<float>(
-          "--phi-deg", [&phi](const float &phi_deg) { phi = phi_deg / 180.f * std::numbers::pi; }, "Specify observer's angle phi")
-      ->default_val(0.f);
+          "--phi-deg", [&phi](const float &phi_deg) { phi = (phi_deg / 180.f) * std::numbers::pi; },
+          "Specify observer's longitude angle phi (default is 180 degrees, observer alogn negative direction of x-axis)")
+      ->default_val(180.f);
 
   demo_subc
       ->add_option<int>("--antialiasing", samples_per_pixel_edge,
@@ -102,8 +129,6 @@ int main(int argc, char **argv) {
       ->default_val(3);
 
   // Float variable definition from command line
-  std::vector<std::string> definition_strings; // QUESTION what is the purpose of this declaration? In fact I tried commenting it
-                                               // out and no problem occurred
   std::unordered_map<std::string, float> floats_from_cl;
   render_subc->add_option_function<std::vector<std::string>>(
       "--define-float",
@@ -126,8 +151,19 @@ int main(int argc, char **argv) {
       },
       "Define named float variables as name=value");
 
-  // TODO add algorithm switch after merge of branch pathtracing, and remember to implement  sub-swithces (e. g. ambient_color for
-  // point light tracing)
+  // Option to decide which rendering algorithm to use
+  std::string render_mode_str = "flat"; // Default is flat tracing
+  render_subc->add_option("-m,--mode", render_mode_str, "Rendering mode: on/off tracing, flat tracing (default), point light tracing or path tracing")
+      ->check(CLI::IsMember({"onoff", "flat", "point_light", "path"}))
+      ->default_val("flat");
+
+  // Parameters for path tracing
+  int n_rays = 10;
+  int russian_roulette_lim = 3;
+  int max_depth = 5;
+  render_subc->add_option("--n_rays", n_rays, "Specify number of rays scattered at every hit (requires path tracing)")->default_val(10);
+  render_subc->add_option("--roulette", russian_roulette_lim, "Specify ray depth reached before russian roulette starts applying (requires path tracing)")->default_val(3);
+  render_subc->add_option("--max-depth", max_depth, "Specify maximum ray depth (requires path tracing)")->default_val(5);
 
   // -----------------------------------------------------------
   // Command line parsing for pfm2png converter mode
@@ -153,14 +189,26 @@ int main(int argc, char **argv) {
 
   // 2. Fill in HdrImage
   std::unique_ptr<HdrImage> img;
+
   if (*demo_subc) {
     // A. (DEMO) Compute the demo image and save PFM file
-    Transformation observer_transformation =
-        rotation_z(phi) * rotation_y(std::numbers::pi / 2.f - theta) *
-        translation(-VEC_X); // evaluates to (identity * backwards translation) for default values of theta and phi
+    Transformation observer_transformation;
 
+    if (mode_str == "onoff") {
+      observer_transformation =
+        translation(-VEC_X) * rotation_z(phi - std::numbers::pi) * rotation_y(std::numbers::pi / 2.f - theta);
+    } else if (mode_str == "path") {
+      observer_transformation =
+        translation(-3.f*VEC_X) * rotation_z(phi - std::numbers::pi) * rotation_y(std::numbers::pi / 2.f - theta);
+    }
+
+    // Generate the demo image accordingly
     std::cout << "Rendering demo image... " << std::flush;
-    img = make_demo_image(orthogonal, width, height, observer_transformation, samples_per_pixel_edge);
+    if (mode_str == "onoff") {
+      img = make_demo_image_onoff(orthogonal, width, height, distance, observer_transformation, samples_per_pixel_edge);
+    } else if (mode_str == "path") {
+      img = make_demo_image_path(orthogonal, width, height, distance, observer_transformation, samples_per_pixel_edge);
+    }
     std::cout << "Done." << std::endl;
 
     // Save PFM image
@@ -169,10 +217,9 @@ int main(int argc, char **argv) {
   } else if (*render_subc) {
     // B. (RENDERER) Parse input file
     std::ifstream is;
-    try {
-      is.open(source_file_name);
-    } catch (const std::exception &err) {
-      std::cerr << "Error opening input (source) file. " << err.what() << '\n';
+    is.open(source_file_name);
+    if (!is) {
+      std::cerr << "Error opening input (source) file \"" << source_file_name << "\"" << std::endl;
       return EXIT_FAILURE;
     }
     InputStream input_stream(is, source_file_name);
@@ -189,10 +236,21 @@ int main(int argc, char **argv) {
     }
 
     ImageTracer tracer = ImageTracer(std::make_unique<HdrImage>(width, height), scene.camera, samples_per_pixel_edge);
-    PointLightTracer renderer{scene.world, Color(0.1f, 0.f, 0.1f), Color(0.f, 0.f, 0.f)};
+
+    std::shared_ptr<Renderer> renderer;
+    if (render_mode_str == "onoff") {
+        renderer = make_shared<OnOffTracer>(scene.world);
+    } else if (render_mode_str == "flat") {
+        renderer = make_shared<FlatTracer>(scene.world, BLACK);
+    } else if (render_mode_str == "point_light") {
+        renderer = make_shared<PointLightTracer>(scene.world, Color(0.1f, 0.1f, 0.05f), BLACK);
+    } else if (render_mode_str == "path"){
+        auto pcg = std::make_shared<PCG>();
+        renderer = make_shared<PathTracer>(scene.world, pcg, n_rays, russian_roulette_lim, max_depth, BLACK);
+    }
 
     std::cout << "Rendering image in " << source_file_name << "... " << std::flush;
-    tracer.fire_all_rays(renderer);
+    tracer.fire_all_rays([&](const Ray& ray) { return (*renderer)(ray); });
     std::cout << "Done." << std::endl;
 
     // Save PFM image
@@ -227,8 +285,16 @@ int main(int argc, char **argv) {
   return EXIT_SUCCESS;
 }
 
-std::unique_ptr<HdrImage> make_demo_image(bool orthogonal, int width, int height, const Transformation &obs_transformation,
-                                          int samples_per_pixel_edge) {
+
+//-------------------------------------------------------------------
+//-------------- HELPER FUNCTIONS --------------
+//-------------------------------------------------------------------
+
+//------------- OnOff Tracing Demo Image -------------
+
+std::unique_ptr<HdrImage> make_demo_image_onoff(bool orthogonal, int width, int height,
+                                                float distance, const Transformation &obs_transformation, int samples_per_pixel_edge) {
+
   // Initialize ImageTracer
   auto img = std::make_unique<HdrImage>(width, height);
 
@@ -240,7 +306,7 @@ std::unique_ptr<HdrImage> make_demo_image(bool orthogonal, int width, int height
     cam = std::make_shared<OrthogonalCamera>(aspect_ratio, obs_transformation);
   } else {
     // provide default *origin-screen* distance, aspect ratio and observer transformation
-    cam = std::make_shared<PerspectiveCamera>(1.f, aspect_ratio, obs_transformation);
+    cam = std::make_unique<PerspectiveCamera>(distance, aspect_ratio, obs_transformation);
   }
   ImageTracer tracer(std::move(img), cam, samples_per_pixel_edge);
 
@@ -258,24 +324,62 @@ std::unique_ptr<HdrImage> make_demo_image(bool orthogonal, int width, int height
     world->add_object(sphere);
   }
 
-  // COMMENT TO RESTORE THE OLD DEMO
-  auto sphere_material =
-      std::make_shared<Material>(std::make_shared<DiffusiveBRDF>(std::make_shared<UniformPigment>(Color(0.5f, 0.5f, 0.5f))));
-  for (auto sphere : world->objects) {
-    sphere->material = sphere_material;
-  }
-  auto source = std::make_shared<PointLightSource>();
-  world->add_light_source(source);
-  PointLightTracer renderer{world, Color(0.1f, 0.f, 0.1f), Color(0.f, 0.f, 0.f)};
-  tracer.fire_all_rays(renderer);
 
-  // UNCOMMENT TO RESTORE OLD DEMO
-  //   // Perform on/off tracing
-  //   tracer.fire_all_rays([&world](Ray ray) -> Color {
-  //     return world->on_off_trace(ray);
-  //   }); // World::on_off_trace requires three arguments, the first one being the World instance, hence it is not
-  //       // compatible with type RaySolver. A lambda wrapping is therefore needed.
+  // Perform on/off tracing
+  tracer.fire_all_rays(OnOffTracer(world));
 
   // Return demo image
   return std::move(tracer.image);
+}
+
+
+
+//--------------------- Path tracing demo image ---------------------
+
+std::unique_ptr<HdrImage> make_demo_image_path(bool orthogonal, int width, int height, float distance, const Transformation &obs_transformation, int samples_per_pixel_edge) {
+  // 1. Create World
+  std::shared_ptr<World> world = std::make_shared<World>();
+
+  // 2. Define Pigments and Materials
+  auto sky_emission = std::make_shared<UniformPigment>(Color(0.2f, 0.3f, 1.f));
+  auto black_pigment = std::make_shared<UniformPigment>(BLACK);
+  auto sky_material = std::make_shared<Material>(std::make_shared<DiffusiveBRDF>(black_pigment), sky_emission);
+
+  auto ground_pattern = std::make_shared<CheckeredPigment>(Color(0.3f, 0.5f, 0.1f), Color(0.1f, 0.2f, 0.5f), 4);
+  auto ground_material = std::make_shared<Material>(std::make_shared<DiffusiveBRDF>(ground_pattern),
+                                                    std::make_shared<UniformPigment>(Color(0.f, 0.f, 0.f)));
+
+  auto grey_pigment = std::make_shared<UniformPigment>(Color(0.5f, 0.5f, 0.5f));
+  auto sphere_material = std::make_shared<Material>(std::make_shared<SpecularBRDF>(grey_pigment), black_pigment);
+
+  auto red_pigment = std::make_shared<UniformPigment>(Color(0.8f, 0.1f, 0.f));
+  auto sphere2_material = std::make_shared<Material>(std::make_shared<DiffusiveBRDF>(red_pigment), black_pigment);
+
+  // 3. Add objects
+  Transformation sky_transform = scaling({50.f, 50.f, 50.f});
+  world->add_object(std::make_shared<Sphere>(sky_transform, sky_material));
+  world->add_object(std::make_shared<Plane>(translation(Vec(0.f, 0.f, -2.f)), ground_material));
+  world->add_object(std::make_shared<Sphere>(scaling({0.4f, 0.4f, 0.4f}), sphere_material));
+  world->add_object(std::make_shared<Sphere>(translation(Vec(0.f, -1.5f, -2.f)), sphere2_material));
+
+  // 4. Setup camera
+  std::unique_ptr<Camera> camera;
+
+  if (orthogonal) { 
+    camera = std::make_unique<OrthogonalCamera>(static_cast<float>(width) / height, obs_transformation);
+  } else {
+    camera = std::make_unique<PerspectiveCamera>(distance, static_cast<float>(width) / height, obs_transformation);
+  }
+
+  // 5. Render image with Montecarlo path tracing
+  auto pcg = std::make_shared<PCG>();
+  PathTracer path_tracer(world, pcg, 10, 2, 6); // n_rays, roulette limit, max_depth
+
+  // 6. Trace the image
+  auto image = std::make_unique<HdrImage>(width, height);
+  ImageTracer image_tracer(std::move(image), std::move(camera));
+  //image_tracer.fire_all_rays(path_tracer);
+  image_tracer.fire_all_rays(path_tracer);
+
+  return std::move(image_tracer.image);
 }